% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Plotting a PCA panel}
\usage{
plot_pca(expr_dt_path, meta_dt_path, output_dir)
}
\arguments{
\item{expr_dt_path}{A file path of the expression table file}

\item{meta_dt_path}{A file path of the metadata file}

\item{output_dir}{A directory of the output file(s)}
}
\description{
Plotting a PCA panel
}
